interview-agent-flask/requirements.txt
Flask==3.1.1
flask_cors==6.0.1
openai==1.93.0
opencv_python==4.11.0.86
pdfplumber==0.11.7
PyPDF2==3.0.1
python_docx==1.2.0
Requests==2.32.4
ws4py==0.6.0
